Split out getSitePlan
